Add fid by id to restaurant detailed
